requirements.txt
langchain
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv
PyPDF2
fastapi
uvicorn
transformers
